pyproject.toml
[project]
name = "tjm"
version = "0.1.0"
description = "Vision-based desktop automation using UGround visual grounding (arxiv 2504.07981)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyautogui>=0.9.54",
    "Pillow>=10.4.0",
    "requests>=2.31.0",
    "pywin32>=306",
    "openai>=1.0.0",
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "transformers>=4.45.0",
    "accelerate>=0.26.0",
    "qwen-vl-utils>=0.0.8",
    "opencv-python>=4.8.0",
    "python-dotenv>=1.0.0",
    "pytest>=8.0.0",
]

[project.scripts]
tjm = "main:main"

[tool.uv.sources]
torch = { index = "pytorch-cu124" }
torchvision = { index = "pytorch-cu124" }

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
